requirements.txt
channels==4.1.0
cn2an==0.5.22
Django==5.0.7
einops==0.8.0
fastapi==0.111.1
faster_whisper==1.0.3
ffmpeg_python==0.2.0
funasr==1.0.0
g2p_en==2.1.0
gradio==4.39.0
gruut==2.3.4
jieba_fast==0.53
LangSegment==0.3.3
LangSegment==0.3.3
librosa==0.9.2
matplotlib==3.9.1
modelscope==1.10.0
nltk==3.8.1
onnxruntime==1.16.3
openai==1.37.1
pandas==2.2.2
psutil==6.0.0
pydub==0.25.1
pyopenjtalk==0.3.3
pypinyin==0.51.0
PyQt5==5.15.11
PyQt5_sip
pytorch_lightning==2.3.1
PyYAML==6.0.1
regex==2024.5.15
Requests==2.32.3
scipy==1.14.0
simpleaudio==1.0.4
soundfile==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.42.3
typeguard==4.3.0
uvicorn==0.30.3
whisper==1.1.10
wordsegment==1.3.1
